move component folder into view
