+++ src/view/component/atom/LoadingMessage/LoadingMessage.tsx
