+++ src/view/component/atom/LoadingMessage/index.ts
