+++ src/view/component/molecule/ExchangeInfoBoard/ExchangeInfoBoard.tsx
@@ -1,4 +1,4 @@
-import { ExchangeInfo } from "../../../model/exchange";
+import { ExchangeInfo } from "../../../../model/exchange";
 
 interface Props {
   exchangeInfo: ExchangeInfo;
