+++ src/view/component/molecule/ExchangeInfoBoard/index.ts
