+++ src/view/component/page/ExchangeInfoPage/ExchangeInfoPage.tsx
@@ -3,16 +3,16 @@ import {
   CurrencyCode,
   TextCurrencyFormatModel,
   InputCurrencyFormatModel,
-} from "../../../model/currency";
-import { ExchangeInfoModel } from "../../../model/exchange";
+} from "../../../../model/currency";
+import { ExchangeInfoModel } from "../../../../model/exchange";
 import {
   MemoryCurrencyFormatRepository,
   useCurrencyFormatViewModel,
-} from "../../../viewmodel/currency";
+} from "../../../../viewmodel/currency";
 import {
   HTTPExchangeInfoRepository,
   useExchangeViewModel,
-} from "../../../viewmodel/exchange";
+} from "../../../../viewmodel/exchange";
 import { ExchangeInfoTemplate } from "../../template/ExchangeInfoTemplate";
 
 interface Props {
