+++ src/view/component/page/ExchangeInfoPage/index.ts
