+++ src/view/component/template/ExchangeInfoTemplate/index.ts
